refactor(prune): extract domain logic from commands/prune to src/prune module
- Move `prune_nodes`, `get_edge_num_muts`, `collapse_sparse_edges_from_leaf_recursive` and helpers to new `src/prune/prune.rs` domain module
- Reduce `commands/prune/run.rs` to thin CLI orchestrator: parse args, load tree, delegate to `crate::prune`, write output
- Relocate domain tests from `commands/prune/__tests__/` to `src/prune/__tests__/`
