--- packages/treetime/src/commands/prune/__tests__/mod.rs
@@ -1,1 +1,0 @@
-mod test_run;
