+++ packages/treetime/src/commands/prune/mod.rs
@@ -1,4 +1,2 @@
-#[cfg(test)]
-mod __tests__;
 pub mod args;
 pub mod run;
