+++ packages/treetime/src/prune/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(test)]
+mod __tests__;
+pub mod prune;
